# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  username: string;
  avatar: string;
  skillLevel: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
}

export interface UpdateProfileInput {
  /** @minLength 1 */
  username: string;
  avatar?: string;
  skillLevel?: string;
}

export interface Player {
  name: string;
  avatar: string;
  level: string;
  confirmed?: boolean;
}

export interface Match {
  id: string;
  title: string;
  hostName: string;
  hostAvatar: string;
  date: string;
  time: string;
  location: string;
  court: string;
  skillLevel: string;
  playersJoined: number;
  playersNeeded: number;
  fee: number;
  status: string;
  description: string;
  players: Player[];
}

export interface CreatePostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  hostName: string;
  date: string;
  time: string;
  location: string;
  court: string;
  skillLevel: string;
  /** @minimum 2 */
  playersNeeded: number;
  fee: number;
  description: string;
}

export interface JoinInput {
  /** @minLength 1 */
  username: string;
}

export interface AcceptInput {
  /** @minLength 1 */
  username: string;
}

export interface ConfirmInput {
  /** @minLength 1 */
  username: string;
}

export interface Product {
  id: string;
  name: string;
  price: number;
  originalPrice?: number;
  condition: string;
  location: string;
  sellerName: string;
  sellerAvatar: string;
  image: string;
  category: string;
  isNewArrival?: boolean;
  isOnSale?: boolean;
}

export interface CreateProductInput {
  /** @minLength 1 */
  name: string;
  price: number;
  condition: string;
  location: string;
  /** @minLength 1 */
  sellerName: string;
  sellerAvatar: string;
  image: string;
  category: string;
}

export interface Clan {
  id: string;
  name: string;
  members: number;
  location: string;
  level: string;
  description: string;
  image: string;
  tags: string[];
  memberNames: string[];
}

export interface CreateClanInput {
  /** @minLength 1 */
  name: string;
  location: string;
  level: string;
  description: string;
  image: string;
  tags: string[];
}