# lib/api-zod/src/generated/types/clan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Clan {
  id: string;
  name: string;
  members: number;
  location: string;
  level: string;
  description: string;
  image: string;
  tags: string[];
  memberNames: string[];
}