# lib/api-zod/src/generated/types/createClanInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClanInput {
  /** @minLength 1 */
  name: string;
  location: string;
  level: string;
  description: string;
  image: string;
  tags: string[];
}