# lib/api-zod/src/generated/types/createPostInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  hostName: string;
  date: string;
  time: string;
  location: string;
  court: string;
  skillLevel: string;
  /** @minimum 2 */
  playersNeeded: number;
  fee: number;
  description: string;
}