# lib/api-zod/src/generated/types/createProductInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductInput {
  /** @minLength 1 */
  name: string;
  price: number;
  condition: string;
  location: string;
  /** @minLength 1 */
  sellerName: string;
  sellerAvatar: string;
  image: string;
  category: string;
}