# lib/api-zod/src/generated/types/joinInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JoinInput {
  /** @minLength 1 */
  username: string;
}