# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RallyHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  name: string;
  avatar: string;
  level: string;
  confirmed?: boolean;
}